chore(types): fixed duplicated types
