+++ app/(opentelemetry-hub-routes)/comparisons/[...slug]/page.tsx
@@ -2,9 +2,9 @@ import 'css/prism.css'
 import 'katex/dist/katex.css'
 
 import { components } from '@/components/MDXComponents'
+import type { Author as Authors } from '@/types/content'
 import { coreContent } from '@/utils/contentlayer/contentUtils'
 import { getAllAuthors } from '@/utils/contentlayer/authorCollection'
-import type { Author as Authors } from '@/utils/contentlayer/authorCollection'
 import OpenTelemetryLayout from '@/layouts/OpenTelemetryLayout'
 import OpenTelemetryHubContent from '@/layouts/OpenTelemetryHubLayout'
 import ComparisonsLayout from '@/layouts/ComparisonsLayout'
