+++ app/(opentelemetry-hub-routes)/guides/[...slug]/page.tsx
@@ -1,9 +1,8 @@
 import 'css/prism.css'
 import { components } from '@/components/MDXComponents'
 import { MDXLayoutRenderer } from 'pliny/mdx-components'
+import type { Guide, Author as Authors } from '@/types/content'
 import { sortPosts, coreContent, allCoreContent } from '@/utils/contentlayer/contentUtils'
-import type { Guide } from '@/utils/contentlayer/guideCollection'
-import type { Author as Authors } from '@/utils/contentlayer/authorCollection'
 import OpenTelemetryLayout from '@/layouts/OpenTelemetryLayout'
 import OpenTelemetryHubContent from '@/layouts/OpenTelemetryHubLayout'
 import GuidesLayout from '@/layouts/GuidesLayout'
