+++ app/(opentelemetry-hub-routes)/opentelemetry/[...slug]/page.tsx
@@ -17,7 +17,7 @@ import { fetchMDXContentByPath, MDXContent } from '@/utils/strapi'
 import { generateStructuredData, safeJsonLdStringify } from '@/utils/structuredData'
 import { compileMDX, MDXRemoteProps } from 'next-mdx-remote/rsc'
 import readingTime from 'reading-time'
-import { CoreContent } from '@/utils/contentlayer/contentUtils'
+import type { CoreContent } from '@/types/content'
 import { mdxOptions, generateTOC } from '@/utils/mdxUtils'
 import { CMS_REVALIDATE_INTERVAL } from '@/constants/cache'
