+++ app/tags/[tag]/page.tsx
@@ -1,9 +1,8 @@
 import { slug } from 'github-slugger'
-import type { CoreContent } from '@/utils/contentlayer/contentUtils'
-import type { Blog } from '@/utils/contentlayer/blogCollection'
+import type { CoreContent, Blog, BlogMeta } from '@/types/content'
 import siteMetadata from '@/data/siteMetadata'
 import ListLayout from '@/layouts/ListLayoutWithTags'
-import { getAllBlogsMeta, type BlogMeta } from '@/utils/contentlayer/blogCollection'
+import { getAllBlogsMeta } from '@/utils/contentlayer/blogCollection'
 import tagData from 'app/tag-data.json'
 import { genPageMetadata } from 'app/seo'
 import { Metadata } from 'next'
