+++ layouts/ComparisonsLayout.tsx
@@ -1,6 +1,5 @@
 import { ReactNode } from 'react'
-import { CoreContent } from '@/utils/contentlayer/contentUtils'
-import type { Author as Authors } from '@/utils/contentlayer/authorCollection'
+import type { CoreContent, Author as Authors } from '@/types/content'
 import ArticleLayout, { TocItemProps } from './ArticleLayout'
 import { RegionProvider } from '@/components/Region/RegionContext'
 import type { Comparison } from '../types/transformedContent'
