+++ layouts/ListLayoutWithTags.tsx
@@ -4,10 +4,7 @@
 import { usePathname } from 'next/navigation'
 import { slug } from 'github-slugger'
 import { formatDate } from 'pliny/utils/formatDate'
-import { CoreContent } from '@/utils/contentlayer/contentUtils'
-import type { Blog } from '@/utils/contentlayer/blogCollection'
-import type { Doc } from '@/utils/contentlayer/docsCollection'
-import type { Guide } from '@/utils/contentlayer/guideCollection'
+import type { CoreContent, Blog, Doc, Guide } from '@/types/content'
 import Link from '@/components/Link'
 import Tag from '@/components/Tag'
 import Button from '@/components/Button/Button'
