+++ layouts/open-telemetry-hub/types.ts
@@ -1,8 +1,5 @@
 import type { ReactNode } from 'react'
-import type { CoreContent } from '@/utils/contentlayer/contentUtils'
-import type { Blog } from '@/utils/contentlayer/blogCollection'
-import type { Author as Authors } from '@/utils/contentlayer/authorCollection'
-import type { Guide } from '@/utils/contentlayer/guideCollection'
+import type { CoreContent, Blog, Author as Authors, Guide } from '@/types/content'
 import type { Comparison } from '../../types/transformedContent'
 import type { MDXContent } from '@/utils/strapi'
