+++ tsconfig.json
@@ -25,7 +25,8 @@
       "@/css/*": ["css/*"],
       "@/api/*": ["api/*"],
       "@/hooks/*": ["hooks/*"],
-      "@/public/*": ["public/*"]
+      "@/public/*": ["public/*"],
+      "@/types/*": ["types/*"]
     },
     "plugins": [
       {
